Extract parse_date in its own module
